+++ TimesynqServer.Application/Service/IUserService.cs
@@ -3,7 +3,7 @@ using TimesynqServer.Application.DTO;
 using TimesynqServer.Application.Pagination;
 using TimesynqServer.Common.Result;
 
-namespace TimesynqServer.Application.Service.UserService
+namespace TimesynqServer.Application.Service
 {
     public interface IUserService
     {
