+++ TimesynqServer.Infrastructure/Service/FollowService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using TimesynqServer.Application.DTO;
 using TimesynqServer.Application.Pagination;
+using TimesynqServer.Application.Service;
 using TimesynqServer.Common;
 using TimesynqServer.Common.Result;
 using TimesynqServer.Domain.Entities;
@@ -8,7 +9,7 @@ using TimesynqServer.Persistence.Projections;
 using TimesynqServer.Persistence.Repository.FollowRepository;
 using TimesynqServer.Persistence.Repository.UserRepository;
 
-namespace TimesynqServer.Application.Service.FollowService
+namespace TimesynqServer.Infrastructure.Service.FollowService
 {
     public class FollowService : IFollowService
     {
