+++ TimesynqServer.Infrastructure/TimesynqServer.Infrastructure.csproj
@@ -18,6 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\TimesynqServer.Application\TimesynqServer.Application.csproj" />
     <ProjectReference Include="..\TimesynqServer.Domain\TimesynqServer.Domain.csproj" />
     <ProjectReference Include="..\TimesynqServer.Persistence\TimesynqServer.Persistence.csproj" />
   </ItemGroup>
