+++ TimesynqServer/Controllers/UserController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using TimesynqServer.Application.DTO;
 using TimesynqServer.Application.Pagination;
-using TimesynqServer.Application.Service.UserService;
+using TimesynqServer.Application.Service;
 using TimesynqServer.Common;
 using TimesynqServer.Common.Result;
 using TimesynqServer.DTO.Request.User;
